fix: type for rsc and rcc are different
