+++ src/index.ts
@@ -1,9 +1,10 @@
 export { MDXRemote } from './rcc'
-export { MDXRemote as MDXRemoteRSC, compileMDX } from './rsc'
+export { MDXRemoteRSC, compileMDX } from './rsc'
 export { serialize } from './serialize'
 export type {
   SerializeOptions,
   MDXRemoteSerializeResult,
   CompileMDXResult,
   MDXRemoteProps,
+  MDXRemoteRSCProps,
 } from './types'
