+++ src/rcc.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState, useMemo } from 'react'
 import { jsxRuntime } from './jsx-runtime'
 import * as mdx from '@mdx-js/react'
 import './idle-callback-polyfill'
-import type { MDXRemoteSerializeResult } from './types'
+import type { MDXRemoteProps } from './types'
 
 // requestIdleCallback types found here: https://github.com/microsoft/TypeScript/issues/21309
 type RequestIdleCallbackHandle = number
@@ -23,25 +23,6 @@ declare global {
     cancelIdleCallback: (handle: RequestIdleCallbackHandle) => void
   }
 }
-
-export type MDXRemoteProps<
-  TScope = Record<string, unknown>,
-  TFrontmatter = Record<string, unknown>,
-> = MDXRemoteSerializeResult<TScope, TFrontmatter> & {
-  /**
-   * A object mapping names to React components.
-   * The key used will be the name accessible to
-   *
-   * For example: `{ ComponentName: Component }` will be accessible in the MDX as `<ComponentName/>`.
-   */
-  components?: React.ComponentProps<typeof mdx.MDXProvider>['components']
-  /**
-   * Determines whether or not the content should be hydrated asynchronously, or "lazily"
-   */
-  lazy?: boolean
-}
-
-export { MDXRemoteSerializeResult }
 
 /**
  * Renders compiled source from serialize.
